Consolidate E2E testing into unified workspace
- Move Playwright dependencies from e2e/package.json to root package.json
- Add yarn scripts: test:e2e and test:e2e:install for unified workspace
- Relocate E2E files: e2e/tests/ → tests/, e2e/lib/ → tests/lib/
- Move playwright.config.ts to project root
- Update import paths in test files to use ./lib/ instead of ../lib/
- Simplify CI workflow to use single workspace commands
- Remove separate e2e workspace (e2e/ directory deleted)
- Update CLAUDE.md documentation to reflect unified workspace structure

Benefits:
- Single Node.js 22.18.0 environment for all development
- Unified dependency management with single yarn.lock
- Simpler CI workflow without workspace switching
- Better IDE integration and development experience
- No more 'cd e2e &&' commands required
