+++ .github/workflows/ci.yml
@@ -23,17 +23,14 @@ jobs:
     steps:
     - uses: actions/checkout@v4
     
-    # Setup Node.js 22 for both app and E2E tests
+    # Setup Node.js 22
     - name: Setup Node.js
       uses: actions/setup-node@v4
       with:
         node-version-file: '.nvmrc'
         cache: 'yarn'
-        cache-dependency-path: |
-          yarn.lock
-          e2e/yarn.lock
     
-    - name: Install main app dependencies
+    - name: Install dependencies
       run: yarn install
     
     - name: Build main app
@@ -50,24 +47,18 @@ jobs:
     - name: Wait for server
       run: npx wait-on http://localhost:3000
     
-    - name: Install E2E dependencies
-      working-directory: ./e2e
-      run: yarn install
-    
     - name: Install Playwright Browsers
-      working-directory: ./e2e
-      run: yarn playwright install --with-deps
+      run: yarn test:e2e:install --with-deps
     
     - name: Wait for Firebase emulator
       run: npx wait-on http://localhost:9099
     
     - name: Run Playwright tests
-      working-directory: ./e2e
-      run: yarn playwright test
+      run: yarn test:e2e
     
     - uses: actions/upload-artifact@v4
       if: ${{ !cancelled() }}
       with:
         name: playwright-report
-        path: e2e/playwright-report/
+        path: playwright-report/
         retention-days: 30
