--- e2e/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "e2e",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "devDependencies": {
-    "@playwright/test": "^1.55.0",
-    "@types/node": "^24.3.0"
-  },
-  "scripts": {
-    "test": "playwright test"
-  }
-}
