+++ tests/lib/VoteAdminTester.ts
