+++ tests/lib/VoteAudienceTester.ts
