+++ tests/lib/VotePresentationTester.ts
