+++ tests/lib/namespace.ts
